7903307: Move ASM dependent classes to a different package
