+++ src/classes/com/sun/tdk/jcov/Instr2.java
@@ -25,8 +25,8 @@
 package com.sun.tdk.jcov;
 
 import com.sun.tdk.jcov.insert.AbstractUniversalInstrumenter;
-import com.sun.tdk.jcov.instrument.ClassMorph;
-import com.sun.tdk.jcov.instrument.ClassMorph2;
+import com.sun.tdk.jcov.instrument.asm.ClassMorph;
+import com.sun.tdk.jcov.instrument.asm.ClassMorph2;
 import com.sun.tdk.jcov.instrument.InstrumentationOptions;
 import com.sun.tdk.jcov.instrument.InstrumentationParams;
 import com.sun.tdk.jcov.tools.EnvHandler;
