+++ src/classes/com/sun/tdk/jcov/instrument/Constants.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2022 Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -31,7 +31,7 @@ package com.sun.tdk.jcov.instrument;
  */
 public class Constants {
 
-    static final String opcNames[] = {
+    public static final String[] opcNames = {
         "nop",
         "aconst_null",
         "iconst_m1",
