+++ src/classes/com/sun/tdk/jcov/instrument/DataAbstract.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2022 Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -96,7 +96,7 @@ public abstract class DataAbstract {
         public int end;
     }
 
-    void xmlGen(XmlContext ctx) {
+    public void xmlGen(XmlContext ctx) {
         xmlTagOpen(ctx, kind());
         ctx.incIndent();
         xmlBody(ctx);
