+++ src/classes/com/sun/tdk/jcov/instrument/DataBlockMethEnter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2022 Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -34,7 +34,7 @@ import java.io.IOException;
  */
 public class DataBlockMethEnter extends DataBlock {
 
-    DataBlockMethEnter(int rootId) {
+    public DataBlockMethEnter(int rootId) {
         super(rootId);
     }
 
@@ -47,7 +47,7 @@ public class DataBlockMethEnter extends DataBlock {
      * the first block.
      */
     @Override
-    boolean isFallenInto() {
+    public boolean isFallenInto() {
         return true;
     }
