+++ src/classes/com/sun/tdk/jcov/instrument/DataBlockTargetCase.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2022 Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -39,7 +39,7 @@ public class DataBlockTargetCase extends DataBlockTarget {
     /**
      * Creates a new instance of DataBlockTargetCase
      */
-    DataBlockTargetCase(int rootId, int value) {
+    public DataBlockTargetCase(int rootId, int value) {
         super(rootId);
         this.targetValue = value;
     }
