+++ src/classes/com/sun/tdk/jcov/instrument/DataBlockTargetCond.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2022 Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -36,7 +36,7 @@ public class DataBlockTargetCond extends DataBlockTarget {
 
     private final boolean side;
 
-    DataBlockTargetCond(int rootId, boolean side) {
+    public DataBlockTargetCond(int rootId, boolean side) {
         super(rootId);
         this.side = side;
     }
