+++ src/classes/com/sun/tdk/jcov/instrument/DataBlockTargetDefault.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2022 Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -38,7 +38,7 @@ public class DataBlockTargetDefault extends DataBlockTarget {
     /**
      * Creates a new instance of DataBlockTargetDefault
      */
-    DataBlockTargetDefault(int rootId) {
+    public DataBlockTargetDefault(int rootId) {
         super(rootId);
     }
