+++ src/classes/com/sun/tdk/jcov/instrument/DataBlockTargetGoto.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2022 Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -35,7 +35,7 @@ import java.io.IOException;
  */
 public class DataBlockTargetGoto extends DataBlockTarget {
 
-    DataBlockTargetGoto(int rootId) {
+    public DataBlockTargetGoto(int rootId) {
         super(rootId);
     }
