+++ src/classes/com/sun/tdk/jcov/instrument/DataBranchSwitch.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2022 Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -44,7 +44,7 @@ public class DataBranchSwitch extends DataBranch {
     /**
      * Creates a new instance of DataBranchSwitch
      */
-    DataBranchSwitch(int rootId, int bciStart, int bciEnd, DataBlockTargetDefault blockDefault) {
+    public DataBranchSwitch(int rootId, int bciStart, int bciEnd, DataBlockTargetDefault blockDefault) {
         this(rootId, bciStart, bciEnd);
         this.blockDefault = blockDefault;
     }
