+++ src/classes/com/sun/tdk/jcov/instrument/DataMethodWithBlocks.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2022 Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -25,6 +25,7 @@
 package com.sun.tdk.jcov.instrument;
 
 import com.sun.tdk.jcov.data.Scale;
+import com.sun.tdk.jcov.instrument.asm.CharacterRangeTableAttribute;
 import com.sun.tdk.jcov.tools.DelegateIterator;
 import java.io.DataInput;
 import java.io.DataOutput;
