+++ src/classes/com/sun/tdk/jcov/instrument/LocationAbstract.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2022 Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -59,7 +59,7 @@ public abstract class LocationAbstract extends DataAbstract implements Comparabl
      * XML Generation
      */
     @Override
-    void xmlAttrs(XmlContext ctx) {
+    protected void xmlAttrs(XmlContext ctx) {
         ctx.attr(XmlNames.START, startBCI());
         ctx.attr(XmlNames.END, endBCI());
     }
