+++ src/classes/com/sun/tdk/jcov/instrument/asm/BlockCodeMethodAdapter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2022 Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -22,13 +22,23 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package com.sun.tdk.jcov.instrument;
+package com.sun.tdk.jcov.instrument.asm;
 
-import com.sun.tdk.jcov.instrument.CharacterRangeTableAttribute.CRTEntry;
 import java.util.*;
 
 import static org.objectweb.asm.Opcodes.*;
 
+import com.sun.tdk.jcov.instrument.DataBlock;
+import com.sun.tdk.jcov.instrument.DataBlockFallThrough;
+import com.sun.tdk.jcov.instrument.DataBlockTargetDefault;
+import com.sun.tdk.jcov.instrument.DataBranchCond;
+import com.sun.tdk.jcov.instrument.DataBranchGoto;
+import com.sun.tdk.jcov.instrument.DataBranchSwitch;
+import com.sun.tdk.jcov.instrument.DataExit;
+import com.sun.tdk.jcov.instrument.DataExitSimple;
+import com.sun.tdk.jcov.instrument.DataMethodWithBlocks;
+import com.sun.tdk.jcov.instrument.InstrumentationParams;
+import com.sun.tdk.jcov.instrument.SimpleBasicBlock;
 import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Attribute;
 
@@ -59,7 +69,7 @@ class BlockCodeMethodAdapter extends OffsetRecordingMethodAdapter {
     private SimpleBasicBlock getBB(AbstractInsnNode insn, int startBCI) {
         SimpleBasicBlock bb = insnToBB.get(insn);
         if (bb == null) {
-            bb = new SimpleBasicBlock(method.rootId, startBCI);
+            bb = new SimpleBasicBlock(method.rootId(), startBCI);
             insnToBB.put(insn, bb);
         } else if (startBCI >= 0) {
             bb.setStartBCI(startBCI);
@@ -137,7 +147,7 @@ class BlockCodeMethodAdapter extends OffsetRecordingMethodAdapter {
                         LabelNode insnTrue = jumpInsn.label;
                         int bciFalse = bcis[insnIdx]; // fall-through
 
-                        DataBranchCond branch = new DataBranchCond(method.rootId, bci, bciFalse - 1);
+                        DataBranchCond branch = new DataBranchCond(method.rootId(), bci, bciFalse - 1);
                         /* DataBlockTarget blockTrue = new DataBlockTargetCond(true);
                          DataBlockTarget blockFalse = new DataBlockTargetCond(false);
                          branch.addTarget(blockTrue);
@@ -174,7 +184,7 @@ class BlockCodeMethodAdapter extends OffsetRecordingMethodAdapter {
 
                         // Create the branch information
                         int bciEnd = bcis[insnIdx] - 1; // end of the switch
-                        DataBranchSwitch branch = new DataBranchSwitch(method.rootId, bci, bciEnd, blockDefault);
+                        DataBranchSwitch branch = new DataBranchSwitch(method.rootId(), bci, bciEnd, blockDefault);
                         // branch.addTarget(blockDefault);
                         exits.add(branch);
 
@@ -211,7 +221,7 @@ class BlockCodeMethodAdapter extends OffsetRecordingMethodAdapter {
 
                         // Create the branch information
                         int bciEnd = bcis[insnIdx] - 1; // end of the switch
-                        DataBranchSwitch branch = new DataBranchSwitch(method.rootId, bci, bciEnd, blockDefault);
+                        DataBranchSwitch branch = new DataBranchSwitch(method.rootId(), bci, bciEnd, blockDefault);
                         // branch.addTarget(blockDefault);
                         exits.add(branch);
 
@@ -239,7 +249,7 @@ class BlockCodeMethodAdapter extends OffsetRecordingMethodAdapter {
 
                         // Create origin info, a branch
                         int bciEnd = bcis[insnIdx] - 1;
-                        DataBranchGoto branch = new DataBranchGoto(method.rootId, bci, bciEnd);
+                        DataBranchGoto branch = new DataBranchGoto(method.rootId(), bci, bciEnd);
                         exits.add(branch);
 
                         // Create destination info, a block target
@@ -263,7 +273,7 @@ class BlockCodeMethodAdapter extends OffsetRecordingMethodAdapter {
                     case ARETURN:
                     case RETURN: {
                         int bciNext = bcis[insnIdx];
-                        DataExit exit = new DataExitSimple(method.rootId, bci, bciNext - 1, insn.getOpcode());
+                        DataExit exit = new DataExitSimple(method.rootId(), bci, bciNext - 1, insn.getOpcode());
                         exits.add(exit);
 
                         AbstractInsnNode insnNext = peek(iit);
@@ -318,7 +328,7 @@ class BlockCodeMethodAdapter extends OffsetRecordingMethodAdapter {
                 for (CharacterRangeTableAttribute.CRTEntry entry : method().getCharacterRangeTable().getEntries()) {
                     if (entry.startBCI() == bci) {
 
-                        if ((entry.flags & CRTEntry.CRT_STATEMENT) != 0 /*& newBlock*/) {
+                        if ((entry.flags & CharacterRangeTableAttribute.CRTEntry.CRT_STATEMENT) != 0 /*& newBlock*/) {
                             newBlock = false;
                             if (insnToBB.get(insn) == null) {
                                 //System.out.println("Should add block at: " + bci + " in " + method().name +
@@ -328,7 +338,7 @@ class BlockCodeMethodAdapter extends OffsetRecordingMethodAdapter {
                             }
                         }
                     } else {
-                        if (entry.endBCI() == index && (entry.flags & CRTEntry.CRT_FLOW_TARGET) != 0) {
+                        if (entry.endBCI() == index && (entry.flags & CharacterRangeTableAttribute.CRTEntry.CRT_FLOW_TARGET) != 0) {
                             newBlock = true;
                         }
                     }
