+++ src/classes/com/sun/tdk/jcov/instrument/asm/BranchCodeMethodAdapter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2022 Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -22,13 +22,30 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package com.sun.tdk.jcov.instrument;
+package com.sun.tdk.jcov.instrument.asm;
 
-import com.sun.tdk.jcov.instrument.CharacterRangeTableAttribute.CRTEntry;
+import com.sun.tdk.jcov.instrument.BasicBlock;
 import java.util.*;
 
 import static org.objectweb.asm.Opcodes.*;
 
+import com.sun.tdk.jcov.instrument.Constants;
+import com.sun.tdk.jcov.instrument.DataBlock;
+import com.sun.tdk.jcov.instrument.DataBlockCatch;
+import com.sun.tdk.jcov.instrument.DataBlockFallThrough;
+import com.sun.tdk.jcov.instrument.DataBlockMethEnter;
+import com.sun.tdk.jcov.instrument.DataBlockTarget;
+import com.sun.tdk.jcov.instrument.DataBlockTargetCase;
+import com.sun.tdk.jcov.instrument.DataBlockTargetCond;
+import com.sun.tdk.jcov.instrument.DataBlockTargetDefault;
+import com.sun.tdk.jcov.instrument.DataBlockTargetGoto;
+import com.sun.tdk.jcov.instrument.DataBranchCond;
+import com.sun.tdk.jcov.instrument.DataBranchGoto;
+import com.sun.tdk.jcov.instrument.DataBranchSwitch;
+import com.sun.tdk.jcov.instrument.DataExit;
+import com.sun.tdk.jcov.instrument.DataExitSimple;
+import com.sun.tdk.jcov.instrument.DataMethodWithBlocks;
+import com.sun.tdk.jcov.instrument.InstrumentationParams;
 import org.objectweb.asm.Label;
 import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Attribute;
@@ -49,7 +66,7 @@ class BranchCodeMethodAdapter extends OffsetRecordingMethodAdapter {
     private final InstrumentationParams params;
 
     public BranchCodeMethodAdapter(final MethodVisitor mv,
-            final DataMethodWithBlocks method, InstrumentationParams params) {
+                                   final DataMethodWithBlocks method, InstrumentationParams params) {
         super(new MethodNode(method.getAccess(), method.getName(), method.getVmSignature(), method.getSignature(), method.getExceptions()),
                 method);
         this.nextVisitor = mv;
@@ -62,7 +79,7 @@ class BranchCodeMethodAdapter extends OffsetRecordingMethodAdapter {
     private BasicBlock getBB(AbstractInsnNode insn, int startBCI) {
         BasicBlock bb = insnToBB.get(insn);
         if (bb == null) {
-            bb = new BasicBlock(method.rootId, startBCI);
+            bb = new BasicBlock(method.rootId(), startBCI);
             insnToBB.put(insn, bb);
         } else if (startBCI >= 0) {
             bb.setStartBCI(startBCI);
@@ -147,7 +164,7 @@ class BranchCodeMethodAdapter extends OffsetRecordingMethodAdapter {
                         LabelNode insnTrue = jumpInsn.label;
                         int bciFalse = bcis[insnIdx]; // fall-through
 
-                        DataBranchCond branch = new DataBranchCond(method.rootId, bci, bciFalse - 1);
+                        DataBranchCond branch = new DataBranchCond(method.rootId(), bci, bciFalse - 1);
                         DataBlockTarget blockTrue = new DataBlockTargetCond(branch.rootId(), true);
                         DataBlockTarget blockFalse = new DataBlockTargetCond(branch.rootId(), false);
                         branch.addTarget(blockTrue);
@@ -184,7 +201,7 @@ class BranchCodeMethodAdapter extends OffsetRecordingMethodAdapter {
 
                         // Create the branch information
                         int bciEnd = bcis[insnIdx] - 1; // end of the switch
-                        DataBranchSwitch branch = new DataBranchSwitch(method.rootId, bci, bciEnd, blockDefault);
+                        DataBranchSwitch branch = new DataBranchSwitch(method.rootId(), bci, bciEnd, blockDefault);
                         branch.addTarget(blockDefault);
                         exits.add(branch);
 
@@ -221,7 +238,7 @@ class BranchCodeMethodAdapter extends OffsetRecordingMethodAdapter {
 
                         // Create the branch information
                         int bciEnd = bcis[insnIdx] - 1; // end of the switch
-                        DataBranchSwitch branch = new DataBranchSwitch(method.rootId, bci, bciEnd, blockDefault);
+                        DataBranchSwitch branch = new DataBranchSwitch(method.rootId(), bci, bciEnd, blockDefault);
                         branch.addTarget(blockDefault);
                         exits.add(branch);
 
@@ -249,7 +266,7 @@ class BranchCodeMethodAdapter extends OffsetRecordingMethodAdapter {
 
                         // Create origin info, a branch
                         int bciEnd = bcis[insnIdx] - 1;
-                        DataBranchGoto branch = new DataBranchGoto(method.rootId, bci, bciEnd);
+                        DataBranchGoto branch = new DataBranchGoto(method.rootId(), bci, bciEnd);
                         exits.add(branch);
 
                         // Create destination info, a block target
@@ -273,7 +290,7 @@ class BranchCodeMethodAdapter extends OffsetRecordingMethodAdapter {
                     case ARETURN:
                     case RETURN: {
                         int bciNext = bcis[insnIdx];
-                        DataExit exit = new DataExitSimple(method.rootId, bci, bciNext - 1, insn.getOpcode());
+                        DataExit exit = new DataExitSimple(method.rootId(), bci, bciNext - 1, insn.getOpcode());
                         exits.add(exit);
 
                         AbstractInsnNode insnNext = peek(iit);
@@ -332,7 +349,7 @@ class BranchCodeMethodAdapter extends OffsetRecordingMethodAdapter {
                 for (CharacterRangeTableAttribute.CRTEntry entry : method().getCharacterRangeTable().getEntries()) {
                     if (entry.startBCI() == bci) {
 
-                        if ((entry.flags & CRTEntry.CRT_STATEMENT) != 0 /*& newBlock*/) {
+                        if ((entry.flags & CharacterRangeTableAttribute.CRTEntry.CRT_STATEMENT) != 0 /*& newBlock*/) {
                             newBlock = false;
                             if (insnToBB.get(insn) == null) {
                                 //System.out.println("Should add block at: " + bci + " in " + method().name +
@@ -342,7 +359,7 @@ class BranchCodeMethodAdapter extends OffsetRecordingMethodAdapter {
                             }
                         }
                     } else {
-                        if (entry.endBCI() == index && (entry.flags & CRTEntry.CRT_FLOW_TARGET) != 0) {
+                        if (entry.endBCI() == index && (entry.flags & CharacterRangeTableAttribute.CRTEntry.CRT_FLOW_TARGET) != 0) {
                             newBlock = true;
                         }
                     }
