+++ src/classes/com/sun/tdk/jcov/instrument/asm/OffsetLabel.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2022 Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -22,44 +22,21 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package com.sun.tdk.jcov.instrument;
+package com.sun.tdk.jcov.instrument.asm;
 
-import java.io.IOException;
-import java.io.InputStream;
-import org.objectweb.asm.ClassReader;
 import org.objectweb.asm.Label;
 
 /**
- * OffsetLabelingClassReader
- *
+ * OffsetLabel
  *
  * @author Robert Field
  */
-class OffsetLabelingClassReader extends ClassReader {
+class OffsetLabel extends Label {
 
-    public OffsetLabelingClassReader(byte[] b) {
-        super(b);
-    }
+    int originalOffset;
+    boolean realLabel;
 
-    public OffsetLabelingClassReader(InputStream in) throws IOException {
-        super(in);
-    }
-
-    /**
-     * @Override public void accept(ClassVisitor cv, Attribute[] attrs, int
-     * flags) { super.accept(new MyClassAdapter(cv), attrs, flags); }
-     ***
-     */
-    @Override
-    protected Label readLabel(int offset, Label[] labels) {
-        OffsetLabel label = (OffsetLabel) labels[offset];
-        if (label == null) {
-            for (int i = 0; i < labels.length; ++i) {
-                labels[i] = new OffsetLabel(i);
-            }
-            label = (OffsetLabel) labels[offset];
-        }
-        label.realLabel = true;
-        return label;
+    OffsetLabel(int originalOffset) {
+        this.originalOffset = originalOffset;
     }
 }
