+++ src/classes/com/sun/tdk/jcov/instrument/asm/OverriddenClassWriter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, 2021,  Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2022  Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -22,7 +22,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package com.sun.tdk.jcov.instrument;
+package com.sun.tdk.jcov.instrument.asm;
 
 import com.sun.tdk.jcov.JREInstr;
 import com.sun.tdk.jcov.runtime.PropertyFinder;
