+++ src/classes/com/sun/tdk/jcov/instrument/asm/SavePointsMethodAdapter.java
@@ -22,7 +22,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package com.sun.tdk.jcov.instrument;
+package com.sun.tdk.jcov.instrument.asm;
 
 import com.sun.tdk.jcov.util.Utils;
 import org.objectweb.asm.MethodVisitor;
