+++ src/classes/com/sun/tdk/jcov/instrument/asm/StaticInvokeMethodAdapter.java
@@ -22,8 +22,9 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package com.sun.tdk.jcov.instrument;
+package com.sun.tdk.jcov.instrument.asm;
 
+import com.sun.tdk.jcov.instrument.InstrumentationParams;
 import com.sun.tdk.jcov.util.Utils;
 import java.util.HashMap;
 import java.util.Map;
