+++ src/classes/com/sun/tdk/jcov/instrument/reader/CharacterRangeTableAttributeStAX.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2022 Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -25,8 +25,8 @@
 package com.sun.tdk.jcov.instrument.reader;
 
 import com.sun.tdk.jcov.data.FileFormatException;
-import com.sun.tdk.jcov.instrument.CharacterRangeTableAttribute;
-import com.sun.tdk.jcov.instrument.CharacterRangeTableAttribute.CRTEntry;
+import com.sun.tdk.jcov.instrument.asm.CharacterRangeTableAttribute;
+import com.sun.tdk.jcov.instrument.asm.CharacterRangeTableAttribute.CRTEntry;
 import com.sun.tdk.jcov.instrument.XmlNames;
 import java.util.ArrayList;
 import java.util.List;
