+++ test/unit/com/sun/tdk/jcov/instrument/plugin/jreinstr/TestPlugin.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2021, 2022 Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -24,7 +24,7 @@
  */
 package com.sun.tdk.jcov.instrument.plugin.jreinstr;
 
-import com.sun.tdk.jcov.instrument.InstrumentationPlugin;
+import com.sun.tdk.jcov.instrument.asm.InstrumentationPlugin;
 import org.objectweb.asm.MethodVisitor;
 
 import java.io.IOException;
